refactor(authJian): Move ChatMessage type from server/pocketbase to typesAndInterfaces
